requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
click>=8.0.0
requests>=2.25.0
psycopg2-binary>=2.9.0

# System monitoring
psutil>=5.9.0

# Retry logic
tenacity>=8.0.0

# Audio transcription
youtube-transcript-api>=0.6.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
yt-dlp>=2023.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Optional: for better type checking
mypy>=1.0.0